Fixed all compile errors. Please read the description.
I changed a number of sections where people were using arrays. Arrays are pretty much pointless for our game and it is much better to use lists for most of our purposes. I also changed an int in Disease into an enum. It's bad to have an int specifying a type because then other people have no idea what it does. Also removed unreachable "break;"s in a switch statement that had returns. If you return, then the break is never reached.